Lab_4 done + a bit of changes in reminder bot

+++ University/year 1 semester 2/Lab_4/Functions.h
@@ -2,17 +2,11 @@
 
 //This is header file. It declares the functions
 
-//extern struct result_struct;
-
-float input_accuracy();
-
-//extern result_struct solution(double accuracy);
-
-//extern result_struct verification(double accuracy);
+double input_accuracy();
 
 void oc_file();
 
-void write_in_file(double res, int n, float accuracy);
+void write_in_file(double res, int n, double accuracy);
 
 void read_from_file();
